fix test suite when running outside of ci

+++ spec/integration_helper.rb
@@ -1,8 +1,14 @@
 require 'spec_helper'
-require 'support/ar_setup'
 require 'support/foo'
 
-RSpec.configure do |c|
-  c.before { Foo.delete_all }
-end
+if ENV["TEST_ENV"] == "travis"
+  require 'support/ar_setup'
 
+  RSpec.configure do |c|
+    c.before { Foo.delete_all }
+  end
+else
+  RSpec.configure do |c|
+    c.filter_run_excluding :integration => true
+  end
+end
